Add files via upload
Boiled down this app consists of two parts. A backend and a frontend files. The frontend is GUI made using python lib. Tkinter. Backend side is mostly functions that are being called when one of the buttons is pressed. 
-User can view all items saved in the db. 
-User can search an item in the db by entering either one or multiple parameters and utilizing the search command
-User can manually add an item in the database using for entry's 'name', 'author', 'year', 'isbn'. Each item in the database has an primary key integer id. 
-User can update when selected an item and parameters entered
-User can delete when selected an item 
-User can close the app using the in app button

To run the app just download and execute the bookstore file